+++ src/test/java/de/kastenklicker/secureserverbackuplibrary/BackupTest.java
@@ -2,7 +2,6 @@ package de.kastenklicker.secureserverbackuplibrary;
 
 import de.kastenklicker.secureserverbackuplibrary.upload.NullUploadClient;
 import de.kastenklicker.secureserverbackuplibrary.upload.SFTPClient;
-import de.kastenklicker.secureserverbackuplibrary.upload.UploadClient;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -10,11 +9,12 @@ import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.utility.MountableFile;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.RandomAccessFile;
+import java.nio.file.Files;
 import java.util.ArrayList;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class BackupTest {
 
@@ -63,14 +63,15 @@ public class BackupTest {
         sftpContainer.start();
         
         // The real testing
-        UploadClient uploadClient = new SFTPClient(
-                sftpContainer.getHost(), sftpContainer.getMappedPort(22), 
+        SFTPClient uploadClient = new SFTPClient(
+                sftpContainer.getHost(),
+                sftpContainer.getMappedPort(22), 
                 "foo",
                 "pass",
-                publicHostKey.getPath(), 
+                publicHostKey, 
                 20000,
                 "/upload");
-
+        
         File backup = new Backup(
                 new ArrayList<>(),
                 backupsDirectory,
@@ -88,6 +89,7 @@ public class BackupTest {
         // Check if file was transferred correctly
         File backupUpload = new File("./src/test/resources/" + backup.getName());
         sftpContainer.copyFileFromContainer("/home/foo/upload/" + backup.getName(), backupUpload.getPath());
+        assertEquals(-1L, Files.mismatch(backup.toPath(), backupUpload.toPath()));
         assertTrue(backupUpload.delete());
     }
 
@@ -120,7 +122,11 @@ public class BackupTest {
     }
 
     @AfterEach
-    public void cleanUp() {
+    public void cleanUp() throws FileNotFoundException {
+        
+        if (backupsDirectory.listFiles() == null)
+            throw new FileNotFoundException(backupsDirectory + " is not a directory.");
+        
         for (File file : backupsDirectory.listFiles()) {
             file.delete();
         }
